move ServerProfileInfoService into core
