+++ core/datastore/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/datastore/datastore_module.kt
@@ -4,6 +4,7 @@ import de.tum.informatics.www1.artemis.native_app.core.common.artemis_context.Ar
 import de.tum.informatics.www1.artemis.native_app.core.datastore.impl.AccountServiceImpl
 import de.tum.informatics.www1.artemis.native_app.core.datastore.impl.ArtemisContextProviderImpl
 import de.tum.informatics.www1.artemis.native_app.core.datastore.impl.ServerConfigurationServiceImpl
+import de.tum.informatics.www1.artemis.native_app.core.datastore.impl.ServerProfileInfoServiceImpl
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
 
@@ -11,4 +12,5 @@ val datastoreModule = module {
     single<AccountService> { AccountServiceImpl(androidContext()) }
     single<ServerConfigurationService> { ServerConfigurationServiceImpl(androidContext()) }
     single<ArtemisContextProvider> { ArtemisContextProviderImpl(get(), get()) }
+    single<ServerProfileInfoService> { ServerProfileInfoServiceImpl(get()) }
 }
