+++ core/datastore/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/core/datastore/impl/ServerProfileInfoServiceImpl.kt
@@ -1,10 +1,10 @@
-package de.tum.informatics.www1.artemis.native_app.feature.login.service.network.impl
+package de.tum.informatics.www1.artemis.native_app.core.datastore.impl
 
 import de.tum.informatics.www1.artemis.native_app.core.data.NetworkResponse
 import de.tum.informatics.www1.artemis.native_app.core.data.performNetworkCall
 import de.tum.informatics.www1.artemis.native_app.core.data.service.KtorProvider
 import de.tum.informatics.www1.artemis.native_app.core.model.server_config.ProfileInfo
-import de.tum.informatics.www1.artemis.native_app.feature.login.service.network.ServerProfileInfoService
+import de.tum.informatics.www1.artemis.native_app.core.datastore.ServerProfileInfoService
 import io.ktor.client.call.body
 import io.ktor.client.request.accept
 import io.ktor.client.request.get
