+++ feature/login/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/login/login_module.kt
@@ -8,10 +8,8 @@ import de.tum.informatics.www1.artemis.native_app.feature.login.service.ServerNo
 import de.tum.informatics.www1.artemis.native_app.feature.login.service.impl.PersistentServerNotificationStorageService
 import de.tum.informatics.www1.artemis.native_app.feature.login.service.network.LoginService
 import de.tum.informatics.www1.artemis.native_app.feature.login.service.network.RegisterService
-import de.tum.informatics.www1.artemis.native_app.feature.login.service.network.ServerProfileInfoService
 import de.tum.informatics.www1.artemis.native_app.feature.login.service.network.impl.LoginServiceImpl
 import de.tum.informatics.www1.artemis.native_app.feature.login.service.network.impl.RegisterServiceImpl
-import de.tum.informatics.www1.artemis.native_app.feature.login.service.network.impl.ServerProfileInfoServiceImpl
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.module.dsl.viewModel
 import org.koin.core.module.dsl.viewModelOf
@@ -49,5 +47,4 @@ val loginModule = module {
     single<RegisterService> { RegisterServiceImpl(get()) }
 
     single<LoginService> { LoginServiceImpl(get()) }
-    single<ServerProfileInfoService> { ServerProfileInfoServiceImpl(get()) }
 }
